+++ lib/gooddata/rest/client.rb
@@ -82,13 +82,21 @@ module GoodData
             new_opts[:username] = username
             new_opts[:password] = password
           end
+
+          if username.is_a?(Hash) && username[:cookies]
+            new_opts[:sst_token] = username[:cookies]['GDCAuthSST']
+            new_opts[:cookies] = username[:cookies]
+          end
+
           unless new_opts[:sst_token]
             fail ArgumentError, 'No username specified' if new_opts[:username].nil?
             fail ArgumentError, 'No password specified' if new_opts[:password].nil?
           end
+
           if username.is_a?(Hash) && username.key?(:server)
             new_opts[:server] = username[:server]
           end
+
           client = Client.new(new_opts)
 
           if client
@@ -99,6 +107,10 @@ module GoodData
 
           # HACK: This line assigns class instance # if not done yet
           @@instance = client # rubocop:disable ClassVars
+        end
+
+        def connect_sso(sso)
+          client = Client.new(sso)
           client
         end